+++ client/src/components/PaginationButtons/index.scss
@@ -1,0 +1,33 @@
+.pagination {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+
+
+  width: 100%;
+
+  .pagination__button {
+    background: var(--saukko-main-white);
+    border-radius: 50%;
+    border: none;
+    height: 40px;
+    width: 40px;
+    font-weight: bold;
+    aspect-ratio: 1;
+
+    &.active {
+      color: var(--saukko-main-white);
+      background: var(--saukko-main-black);
+    }
+
+    &:disabled {
+      cursor: not-allowed;
+    }
+
+    &:not(:disabled):not(.active):hover {
+      background-color: rgba(0, 0, 0, 0.1);
+      cursor: pointer;
+    }
+  }
+}
